+++ lib/Command/ThruwayRouterCommand.php
@@ -41,7 +41,7 @@ class ThruwayRouterCommand extends ContainerAwareCommand
      */
     protected function execute( InputInterface $input, OutputInterface $output ): int
     {
-        if ( $this->getParameter( 'vs_thruway' )['enable_logging'] ) {
+        if ( $this->getParameter( 'vs_thruway.enable_logging' ) ) {
             \Thruway\Logging\Logger::set( $this->logger );
         } else {
             \Thruway\Logging\Logger::set( new \Psr\Log\NullLogger() );
@@ -50,14 +50,14 @@ class ThruwayRouterCommand extends ContainerAwareCommand
         try {
             $output->writeln( 'Making a go at starting the Thruway Router' );
 
-            //Configure stuff
-            $config = $this->getParameter( 'vs_thruway' );
-
             //Get the Router Service
             $server = $this->get( 'vs_thruway.thruway.server' );
 
             //Trusted provider (bound to loopback and requires no authentication)
-            $trustedProvider = new RatchetTransportProvider( $config['router']['ip'], $config['router']['trusted_port'] );
+            $trustedProvider = new RatchetTransportProvider(
+                $this->getParameter( 'vs_thruway.router_ip' ),
+                $this->getParameter( 'vs_thruway.router_trusted_port' )
+            );
             $trustedProvider->setTrusted( true );
             $server->addTransportProvider( $trustedProvider );